requirements.txt
ultralytics
Flask
flask-socketio
eventlet
numpy
Pillow
